Adjust #[cfg] etc. to make new <pre> work without CSS.
